refactor: extract loginValidate and signUpValidate to types
