+++ src/hooks/userHooks.ts
@@ -1,12 +1,15 @@
-import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { LoginFormProps, SignUpFormProps, loginType, signUpType } from "types/userTypes";
 
-export const loginValidate = z.object({
-  username: z.string().min(1, "Email obrigatório").email("Email inválido"),
-  password: z.string().min(1, "Senha obrigatória")
-});
+import {
+  LoginFormProps,
+  SignUpFormProps,
+  loginType,
+  loginValidate,
+  signUpType,
+  signUpValidate
+} from "types/userTypes";
+
 
 export function useLogin() {
   const {
@@ -31,17 +34,6 @@ export function useLogin() {
     getFieldState,
   };
 }
-
-export const signUpValidate = z.object({
-  name: z.string().min(1, "Nome é obrigatório"),
-  username: z.string().min(1, "Email é obrigatório").email("Email inválido"),
-  password1: z.string().min(6, "Senha deve ter no mínimo 6 caracteres")
-    .regex(/(?=.*[A-Z])(?=.*[0-9])/, { message: "A senha deve conter letras e números" }),
-  password2: z.string(),
-}).refine((fields) => fields.password1 === fields.password2, {
-  path: ["password2"],
-  message: "Senhas diferentes",
-});
 
 export function useSignUp() {
   const {
